LFS/requirements.txt
boto3
click
rich
python-dotenv
